requirements.txt
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.0.0
plotly>=5.15.0
python-dotenv>=1.0.0
stravalib>=1.4.0
scipy>=1.10.0